Removed package refactor as this would cause a bug.
