+++ app/src/test/java/nl/tudelft/pixelperfect/pixelperfect/clienttest/EventCompletedMessageTest.java
@@ -4,7 +4,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import nl.tudelft.pixelperfect.client.message.EventCompletedMessage;
+import nl.tudelft.pixelperfect.client.EventCompletedMessage;
 
 /**
  * Tests the EventCompleted message.
