+++ app/src/test/java/nl/tudelft/pixelperfect/pixelperfect/clienttest/EventsMessageTest.java
@@ -4,7 +4,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import nl.tudelft.pixelperfect.client.message.EventsMessage;
+import nl.tudelft.pixelperfect.client.EventsMessage;
 
 /**
  * Test case for the event message class.
